refactor(test_utils): utility to handle `NodeMsg`
